Add ramune with tests and move artisanale to my repertory.
